pyproject.toml
[tool.poetry]
name = "document-classifier"
version = "0.1.0"
description = ""
authors = ["Bischoff, Silas <SBischoff@brockhaus-ag.de>"]

[tool.poetry.dependencies]
python = "^3.6"
keras = "^2.2"
pillow = "^6.1"
tensorflow-gpu = "^1.14"
flask = "^1.1"
flask_cors = "^3.0"
flask_yoloapi = "^0.1.6"
numpy = "^1.17"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
